Avoid widgets dealing with users config
PasswordWidget and PublicKeySelector widgets were reading the
Yast::UsersSimple database into their own instance of YUsers::Config and
writing it back when executing the #store method, which might lead to
sync problems.

To avoid it, now those widgets do not deal with the config anymore.
Instead, they just receive the user as a parameter during the
initialization and make a copy to perform needed validations before
modifying it.

For reading and writing from/to Yast::UsersSimple, a new InstRootFirts
client has been added.