Flush removed elements after each write
The Users module was not reseting its internal data structure for
removed elements after each write action.

With previous code writing the /etc/passwd and /etc/groups files
directly, this was not a problem. It simply didn't write the "deleted"
elements. Now that new code is running a "userdel" and "groupdel"
commands to remove them, an error for the first previously deleted
element is prompted after running a "Write Changes Now" action.

Reseting "removed" lists internally at the end of each #Write action
fixes the problem.